Remove tag logic refactor

+++ src/main/java/com/happycoders/settings/SettingsController.java
@@ -188,28 +188,31 @@ public class SettingsController {
     }
 
     //AJAX
-    @PostMapping ("/settings/tags/add")
+    @PostMapping (SETTINGS_TAGS_URL + "/add")
     @ResponseBody
     public ResponseEntity addTag (@CurrentUser Account account, @RequestBody TagForm tagForm) {
         String title = tagForm.getTagTitle();
-        Tag tag = tagRepository.findByTitle(title)
-                .orElseGet(() -> tagRepository.save(Tag.builder()
-                .title(tagForm.getTagTitle())
-                .build()));
-
+//        Tag tag = tagRepository.findByTitle(title)
+//                .orElseGet(() -> tagRepository.save(Tag.builder()
+//                .title(tagForm.getTagTitle())
+//                .build()));
+        Tag tag = tagRepository.findByTitle(title);
+        if (tag == null) {
+            tag = tagRepository.save(Tag.builder().title(title).build());
+        }
         accountService.addTag(account, tag);
-
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping ("/settings/tags/remove")
+    @PostMapping (SETTINGS_TAGS_URL +"/remove")
     @ResponseBody
     public ResponseEntity removeTag (@CurrentUser Account account, @RequestBody TagForm tagForm) {
         String title = tagForm.getTagTitle();
-        Tag tag = tagRepository.findByTitle(title).orElseThrow();
-
+        Tag tag = tagRepository.findByTitle(title);
+        if (tag == null) {
+            return ResponseEntity.badRequest().build();
+        }
         accountService.removeTag(account, tag);
-
         return ResponseEntity.ok().build();
 
     }
